remove scala test case, replaced by junit case. And fixed arithmetic precision issue.

+++ src/main/java/me/raycai/java102/repl/model/Expression.java
@@ -1,5 +1,7 @@
 package me.raycai.java102.repl.model;
 
+import java.math.BigDecimal;
+
 public interface Expression {
-    float evaluate();
+    BigDecimal evaluate();
 }
